requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
bitsandbytes==0.42.0
blinker==1.9.0
blobfile==3.0.0
boto3==1.37.1
botocore==1.37.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
community==1.0.0b1
dataclasses-json==0.6.7
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-aws==0.2.13
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.11
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==2.0.2
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.2
safetensors==0.5.3
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
